refac: Move authentication logic from `app` to package `api`
