+++ packages/api/src/utils/authTokens.ts
@@ -1,10 +1,12 @@
-/* 
+/*
  * Fonctions pour enlever ou ajouter les cookies qui permettent de partager
- * l'authentification entre les sous-domaines. 
+ * l'authentification entre les sous-domaines.
 
  * Ref: https://github.com/orgs/supabase/discussions/5742#discussioncomment-4050444
 */
-import {Session, SupabaseClient} from '@supabase/supabase-js';
+import { Session, SupabaseClient } from '@supabase/supabase-js';
+import { ENV } from '../environmentVariables';
+import { supabaseClient } from './supabase-client';
 
 export const ACCESS_TOKEN = 'tet-access-token';
 export const REFRESH_TOKEN = 'tet-refresh-token';
@@ -27,9 +29,9 @@ const SUPABASE_TOKEN = /^sb-.*-auth-token/;
 const deleteSbCookies = (domain: string) =>
   document.cookie
     .split(/\s*;\s*/)
-    .map(cookie => cookie.split('='))
-    .filter(x => x[0].match(SUPABASE_TOKEN))
-    .forEach(x => (document.cookie = formatExpiredToken(x[0], domain)));
+    .map((cookie) => cookie.split('='))
+    .filter((x) => x[0].match(SUPABASE_TOKEN))
+    .forEach((x) => (document.cookie = formatExpiredToken(x[0], domain)));
 
 /** Crée les tokens à partir de la session */
 const MAX_AGE = 60 * 60 * 24 * 365;
@@ -51,6 +53,24 @@ export const setAuthTokens = (session: Session, domain: string) => {
   );
 };
 
+export async function getSession() {
+  const { data, error } = await supabaseClient.auth.getSession();
+  if (data?.session) {
+    return data.session;
+  }
+  if (error) throw error;
+
+  // restaure une éventuelle session précédente
+  const ret = await restoreSessionFromAuthTokens(supabaseClient);
+  if (ret) {
+    const { data, error } = ret;
+    if (data?.session) {
+      return data.session;
+    }
+    if (error) throw error;
+  }
+}
+
 /** Restaure la session depuis les tokens */
 export const restoreSessionFromAuthTokens = async (
   supabase: SupabaseClient
@@ -62,9 +82,9 @@ export const restoreSessionFromAuthTokens = async (
   // recherche les cookies
   const cookies = document.cookie
     .split(/\s*;\s*/)
-    .map(cookie => cookie.split('='));
-  const accessTokenCookie = cookies.find(x => x[0] === ACCESS_TOKEN);
-  const refreshTokenCookie = cookies.find(x => x[0] === REFRESH_TOKEN);
+    .map((cookie) => cookie.split('='));
+  const accessTokenCookie = cookies.find((x) => x[0] === ACCESS_TOKEN);
+  const refreshTokenCookie = cookies.find((x) => x[0] === REFRESH_TOKEN);
 
   if (accessTokenCookie && refreshTokenCookie) {
     return supabase.auth.setSession({
@@ -74,3 +94,13 @@ export const restoreSessionFromAuthTokens = async (
   }
   return null;
 };
+
+export async function getAuthHeaders() {
+  const session = await getSession();
+  return session?.access_token
+    ? {
+        authorization: `Bearer ${session.access_token}`,
+        apikey: `${ENV.supabase_anon_key}`,
+      }
+    : null;
+}
