+++ packages/api/src/utils/supabase-client.ts
@@ -1,0 +1,13 @@
+import { createClient } from '@supabase/supabase-js';
+import { Database } from '@tet/api';
+import { ENV } from '../environmentVariables';
+
+export const supabaseClient = createClient<Database>(
+  ENV.supabase_url as string,
+  ENV.supabase_anon_key as string,
+  {
+    db: {
+      schema: 'public',
+    },
+  }
+);
